+++ src/ISADotnet/API/Protocol.fs
@@ -3,6 +3,39 @@ namespace ISADotNet.API
 open ISADotNet
 open Update
 
+
+(*type ProtocolParameter with*)
+
+module ProtocolParameter =
+
+    /// Returns the name of the paramater as string if it exists
+    let tryGetNameAsString (pp : ProtocolParameter) =
+        pp.ParameterName
+        |> Option.bind (OntologyAnnotation.tryGetNameAsString)
+
+    /// Returns the name of the paramater as string
+    let getNameAsString (pp : ProtocolParameter) =
+        tryGetNameAsString pp
+        |> Option.defaultValue ""
+
+    /// Returns the name of the paramater and its number as string (e.g. "temperature #2")
+    let getNameAsStringWithNumber (pp : ProtocolParameter) =
+        pp.ParameterName
+        |> Option.map (OntologyAnnotation.getNameAsStringWithNumber)
+        |> Option.defaultValue ""
+        
+    /// Returns true if the given name matches the name of the parameter
+    let nameEqualsString (name : string) (pp : ProtocolParameter) =
+        match pp.ParameterName with
+        | Some oa -> OntologyAnnotation.nameEqualsString name oa
+        | None -> false
+
+    /// Returns true if the given numbered name matches the name of the parameter (e.g. "temperature #2")
+    let nameWithNumberEqualsString (name : string) (pp : ProtocolParameter) =
+        match pp.ParameterName with
+        | Some oa -> OntologyAnnotation.nameWithNumberEqualsString name oa
+        | None -> false
+
 module Protocol =  
   
     ///// If a protocol for which the predicate returns true exists in the study, gets it