+++ src/ISADotnet/Printer.fs
@@ -1,0 +1,6 @@
+﻿namespace ISADotNet
+
+type IISAPrintable =
+    abstract member Print : unit -> string
+
+    abstract member PrintCompact : unit -> string